# sqlbase: don't unset datums unless necessary