Previously, the row fetcher would unset every column in its EncDatumRow
on every new row it received. This was surprisingly inefficient -
restting the fields in an EncDatum requires a memory barrier, which is
slow enough to make a dent in performance when its in a tight loop.

Now, we calculate the set of nullable needed columns at plan time, and
only unset those.

NB: this has the potential to be unsafe if we have bad data on disk. If
there are some rows that are missing non-nullable columns, then the row
fetcher would mistakenly use stale data for those columns.

Is that okay? If not, we could either always unset needed columns regardless of nullability. Or, we could keep a `seen` slice for the nullable, needed columns in the fetch plan, and check it at the end of each row to see if there are any missing ones, returning an error at that point instead of wrong data.

```
name                                old time/op    new time/op    delta
WideTableIgnoreColumns/Cockroach-8    14.3ms ± 1%    12.8ms ± 1%  -10.19%  (p=0.000 n=10+10)

name                                old alloc/op   new alloc/op   delta
WideTableIgnoreColumns/Cockroach-8    2.11MB ± 0%    2.10MB ± 0%   -0.43%  (p=0.000 n=9+9)

name                                old allocs/op  new allocs/op  delta
WideTableIgnoreColumns/Cockroach-8     1.37k ±22%     1.17k ± 3%  -14.73%  (p=0.005 n=10+9)

name               old time/op    new time/op    delta
Count/Cockroach-8    49.8ms ± 2%    49.1ms ± 1%  -1.32%  (p=0.002 n=10+9)
```

Release note (performance improvement): improve the performance of scans
by doing less work