@petermattis 2018-01-19
I'm a little wary of this change as it seems like has the potential for a bug in the presence of schema changes. As an alternative, can we keep track in a `FastIntSet` which columns were set and only clear the ones that were set?

---

Review status: 0 of 1 files reviewed at latest revision, all discussions resolved, all commit checks successful.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/21572#-:-L3DhOjEP09-2QHq2VxF:b-m9a6c2)*
<!-- Sent from Reviewable.io -->