@nvanbenschoten 2018-01-19
@jordanlewis FWIW I'm seeing the following experiment give about a 3-4% speedup:

```
diff --git a/pkg/sql/sqlbase/encoded_datum.go b/pkg/sql/sqlbase/encoded_datum.go
index 3bfae8bc1..fe89fb4b1 100644
--- a/pkg/sql/sqlbase/encoded_datum.go
+++ b/pkg/sql/sqlbase/encoded_datum.go
@@ -422,6 +422,17 @@ func (r EncDatumRow) CompareToDatums(
        return 0, nil
 }

+const emptyDatumRowChunk = 128
+
+var emptyDatumRow = make(EncDatumRow, emptyDatumRowChunk)
+
+// UnsetAllDatums unsets all Datum values in the row.
+func (r EncDatumRow) UnsetAllDatums() {
+       for i := 0; i < len(r); i += emptyDatumRowChunk {
+               copy(r[i:], emptyDatumRow)
+       }
+}
+
 // EncDatumRows is a slice of EncDatumRows having the same schema.
 type EncDatumRows []EncDatumRow

diff --git a/pkg/sql/sqlbase/rowfetcher.go b/pkg/sql/sqlbase/rowfetcher.go
index a817b35cc..2d5f56a85 100644
--- a/pkg/sql/sqlbase/rowfetcher.go
+++ b/pkg/sql/sqlbase/rowfetcher.go
@@ -584,9 +584,7 @@ func (rf *RowFetcher) processKV(

                // Reset the row to nil; it will get filled in with the column
                // values as we decode the key-value pairs for the row.
-               for i := range table.row {
-                       table.row[i].UnsetDatum()
-               }
+               table.row.UnsetAllDatums()

                // Fill in the column values that are part of the index key.
                for i := range table.keyVals {
```

```
name                                old time/op  new time/op  delta
WideTableIgnoreColumns/Cockroach-4  11.4ms ± 1%  11.0ms ± 1%  -3.67%  (p=0.000 n=9+10)
```

It's not as much of an improvement as we're seeing here, but it might be a good alternative if we find this too risky.