@jordanlewis 2018-01-19
Thanks Nathan for the experiment! I want to play around to see if there's a way to unset only the things that we have to. `COUNT(*)` queries for example should never have to zero the memory, since none of the columns in the row are set at all.

I'm about to push a change that just gets rid of the nullability condition, and instead just unconditionally zeroes all needed columns. We could get fancy and have a hybrid approach I guess where if the needed columns form a contiguous block we zero them altogether, but I'm not sure if that's worth trying.