@jordanlewis 2018-01-19
The new numbers:

```
name                                old time/op    new time/op    delta
WideTableIgnoreColumns/Cockroach-8    12.2ms ± 1%    11.4ms ± 4%  -6.33%  (p=0.000 n=9+10)

name                                old alloc/op   new alloc/op   delta
WideTableIgnoreColumns/Cockroach-8    2.10MB ± 0%    2.09MB ± 0%  -0.22%  (p=0.034 n=10+8)

name                                old allocs/op  new allocs/op  delta
WideTableIgnoreColumns/Cockroach-8     1.37k ±25%     1.32k ±29%    ~     (p=0.352 n=10+10)
```