@petermattis 2018-01-19
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

---

Review status: 0 of 1 files reviewed at latest revision, all discussions resolved, some commit checks failed.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/21572#-:-L3EZxblgjwTzW40UV6Z:bnfp4nl)*
<!-- Sent from Reviewable.io -->